Use modern META filename
The `META.info` is a legacy, pre-Christmas name. While it's still currently supported,
`META6.json` is the new name. And since a lot of people simply copy some module's 
structure, the old name still proliferates, so we're trying to get rid of it for good 
by sending PRs to any modules that use the old name, to switch to the modern name.